refactor: extract out payment discount logic
