+++ src/Utils/index.js
@@ -14,3 +14,4 @@ export { getAddress } from './get-address';
 export { addNewAddress } from './add-new-address';
 export { editAddress } from './edit-address';
 export { deleteAddress } from './delete-address';
+export * from './price-details';
